Error/SR messages generation for PersonalDetails moved to util function
